Various changes with gravitational and magnetic field models

+++ Kit/Include/mathkit.h
@@ -69,9 +69,9 @@ void QTxV(double QAB[4], double Va[3], double Vb[3]);
 void UNITQ(double Q[4]);
 void RECTIFYQ(double Q[4]);
 void PerpBasis(double A[3], double B[3], double C[3]);
-long fact(long const n);
-long oddfact(long const n);
-long factDfact(long const n, long const m);
+double fact(long const n);
+double oddfact(long const n);
+double factDfact(long const n, long const m);
 void Legendre(long N, long M, double x, double P[19][19], double sdP[19][19]);
 void SphericalHarmonics(const long N, const long M, const double r,
                         const double trigs[4], const double Re, const double K,
